hardware/mordomo - (orange-pi-5-16gb)/ecossistemas/mordomo/containers/tts/tts-engine/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio processing
numpy==1.24.3
soundfile==0.12.1
scipy==1.11.4

# Local TTS models
piper-tts==1.2.0
TTS==0.22.0  # Coqui TTS
# edge-tts==6.1.9  # Microsoft Edge TTS (API-based but free)

# API-based TTS
openai==1.3.7
azure-cognitiveservices-speech==1.32.1

# NATS for event bus
nats-py==2.6.0

# Testing and monitoring
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
